+++ src/applications/modules/bird_watching/core/bird_utils.h
@@ -7,15 +7,14 @@ namespace BirdWatching {
 namespace Utils {
 
 /**
- * @brief 智能检测小鸟的帧数
- * 
- * 使用优化的查找算法，先尝试常见帧数，再精确查找
- * 避免扫描整个目录，大幅提升检测速度
- * 
+ * @brief 从bundle文件检测小鸟的帧数
+ *
+ * 直接从bundle.bin文件头读取帧数，O(1)时间复杂度
+ *
  * @param bird_id 小鸟ID
- * @return 检测到的帧数，如果没有找到则返回0
+ * @return 检测到的帧数（最大65535），如果没有找到则返回0
  */
-uint8_t detectFrameCount(uint16_t bird_id);
+uint16_t detectFrameCount(uint16_t bird_id);
 
 } // namespace Utils
 } // namespace BirdWatching
